added jekyll, added basic framework, test post
I've installled Jekyll with the gh-pages gem, and configured it a
little. I have also started to build a simple framework for posting,
with header and footer snippets and one initial layout.

I have also created a test post for, funnily, testing. I have tried to
create a main page indexing the posts, but have yet to be successful.