pagination config, index.html, new home template
I have changed the configuration for pagination in the _config.yml file.
This was because I had blindly copied a snipped from the docs that
didn't actually match my setup.

As well as this, I added a new template for the homepage, and changed
the index.html file a load. I had essentially misunderstood pagination
quite a lot. We're on a route to success now, m8.