css, another test post, includes, config, index,..
A stylesheet has been added, and I've started to create an actual
layout. I've also changed the name of the blog (say hello to h tea m
l), which I'll likely dislike in a week. I also changed the header and
footer includes, as well as the home and post layouts, as part of
creating the layout.

I added a "subname" variable to the config file too, meaning I can
access a strapline, slogan type thing wherever. It may not be mentioned
in the Jekyll docs, but it's perfectly all right if you ask me.

As part of getting Pagination to work, which still isn't, I created
another quick test post. index.html has changed as well in my endeavour
to paginate, but I think I spy my error already.