pagination, index, theming
I have changed the configuration file to add a slightly longer
description, remove the spaces from the website name, and increase the
pagination post count to 12.

I have developed the layout and theme, so that it is almost complete. I
still need to sort out an issue with the footer and main section
overlap, and perhaps give links, code, etc. some interesting styles. The
header include has changed to allow for the changes in the header bar
and the moving of assets, as well as to wrap the content in another div.
The footer include has changed accordingly.

I have moved css and images to a new assets directory, and also brought
in Montserrat Hairline in WOFF format. For whatever reason, this is not
available on Google Fonts.