first post, layouts, styles, sprite, favicon,..
One of the most significant changes in my stupid mind is the addition of
the first post, complete. I have not proofread it. There is a low chance
I ever will.

In news relating to real changes, I have made a lot. I have, in fact,
neglected making this commit as the fist post was half-done for a while,
and I didn't want to commit it without it being finished.

I have modified the header and footer includes. For one thing, I have
started using a main tag, rather than a section, despite it's queer level
of browser support. I have also added a script element calling the
colourtime.js script mentioned later.

The two layouts have been changed. I completed the pagination business,
and moved the page indicator to the home layout rather than the index
file purely as it felt more correct. It's something common across all of
the home pages, so I feel like it should be in the home template. The
post template has been changed, removing the silly comment form and
adding a little profile of me after the post. There is also a nicely
formatted date below the post title.

I have removed the two test posts, and added the first real one, as
partly mentioned.

I have added the colourtime.js script, which slowly and gradually changes
the colour of a few elements of the page. This uses the current time and
some sine wave calculations to create nice, fairly randomised fades
between different colours. The idea is to load the page and see it
displaying one colour, read a post, look at the header (for
instance) once more, and then question one's sanity.

A sprite in both PNG and SVG format has been added, and integrated. This
accounts for the main logo, and the social links in the profile box
after posts. As I have started using this sprite for the main logo, I
have removed the Montserrat Hairline WOFF from the assets and
stylesheet.

I have added a favicon, too. I decided to steal the one I made for
functionalstoneware.com, as it's fitting enough for both.

I'll commit more often now.